fix(rest-dsl): Cannot parse Apicurio OpenApi
Currently, the `summary` field from Apicurio-generated OpenApi cannot
be parsed.

This commit adds a check to determine whether the operation is an object
or a string before trying to parse it.

fixes: https://github.com/KaotoIO/kaoto-ui/issues/2279
